Tightened up variable-case facets
