+++ int-test/src/test/scala/au/csiro/data61/magda/api/FacetSpec.scala
@@ -136,25 +136,6 @@ class FacetSpec extends BaseSearchApiSpec {
 
       def getFacet(result: SearchResult) = result.facets.get.find(_.id.equals(facetType.id)).get
 
-      it("(meta) filter and groupResult should line up") {
-        checkFacetsNoQuery { (dataSets: List[DataSet], facetSize: Int) ⇒
-          val grouped = groupResult(dataSets)
-          val reduced = dataSets.map(reducer).flatten
-
-          whenever(!reduced.isEmpty) {
-            reduced.foreach { value =>
-              val filtered = dataSets.filter(filter(_, FacetOption(value = value, hitCount = 0))).toSet
-              withClue(s"with option $key, grouped grouped and datasets ${dataSets.map(reducer)}") {
-                grouped.get(value) match {
-                  case Some(group) => group should equal(filtered)
-                  case None        => filtered.size should equal(0)
-                }
-              }
-            }
-          }
-        }
-      }
-
       describe("all facet options should correspond with grouping the datasets for that query") {
         it("without query") {
           checkFacetsNoQuery { (dataSets: List[DataSet], facetSize: Int) ⇒
@@ -186,11 +167,11 @@ class FacetSpec extends BaseSearchApiSpec {
 
               val matched = facet.options.filter(_.matched)
               whenever(matched.size > 0 && outerResult.strategy.get == MatchAll) {
+                val groupedResults = groupResult(dataSets).mapValues(_.size)
                 matched.foreach { option ⇒
-                  val facetDataSets = dataSets.filter(filter(_, option))
 
-                  withClue(s"For option ${option} and grouped datasets ${groupResult(dataSets).mapValues(_.size)} and all options ${facet.options}") {
-                    facetDataSets.size.toLong should equal(option.hitCount)
+                  withClue(s"For option ${option} and grouped datasets ${groupedResults} and all options ${facet.options}") {
+                    groupedResults(option.value).toLong should equal(option.hitCount)
                   }
                 }
               }
